requirements.txt
Cython>=0.21.1
ftfy>=4.1.0,<5.0.0
lxml
numpy>=1.11.0
pytest>=4.0.0
pytest-cov>=2.6.0
scikit-learn>=0.22.0
scipy>=0.17.0
sklearn-crfsuite==0.3.6
dateparser==1.0.0
joblib==0.17.0
catboost==0.24.2